chore(refactor): vite - extract default vite config
