+++ packages/vite/src/lib/envVarDefinitions.ts
